+++ sync-for-commerce/src/codatsynccommerce/models/operations/list_companies.py
@@ -26,9 +26,12 @@ class ListCompaniesRequest:
 @dataclasses.dataclass
 class ListCompaniesResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     companies: Optional[shared_companies.Companies] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""