+++ sync-for-commerce/src/codatsynccommerce/models/operations/list_integrations.py
@@ -26,9 +26,12 @@ class ListIntegrationsRequest:
 @dataclasses.dataclass
 class ListIntegrationsResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     integrations: Optional[shared_integrations.Integrations] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""