+++ sync-for-commerce/src/codatsynccommerce/models/shared/localization.py
@@ -12,6 +12,7 @@ from typing import Optional
 @dataclasses.dataclass
 class Localization:
     required: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('required'), 'exclude': lambda f: f is None }})
+    r"""`True` if required."""
     text: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('text'), 'exclude': lambda f: f is None }})
     r"""Value of the property."""